Added getting products, Added converting byte array to image demo

+++ VendingMachine.Database/Producten.sql
@@ -2,6 +2,6 @@
 (
 	[Id] INT IDENTITY(1,1) NOT NULL PRIMARY KEY, 
     [Naam] NVARCHAR(50) NOT NULL, 
-    [Prijs] DECIMAL(5) NOT NULL, 
+    [Prijs] FLOAT NOT NULL, 
     [Afbeelding] IMAGE NULL
 )
